refactor: implement suggestions
