+++ src/jwt/verify.ts
@@ -5,31 +5,27 @@ import {
   decodeFromBase64UrlToBuffer,
   importKey,
   textEncoder,
-} from "./utils";
-import type { JWTRegisteredClaims } from "./types";
+} from "./_utils";
+import type { JWTRegisteredClaims, VerifyJWTOptions } from "./types";
 // @ts-expect-error: will resolve at runtime
 import { subtle } from "uncrypto";
 
-export async function verifyJWT({
-  token,
-  secret,
-  issuer,
-  audience,
-  leeway = 60,
-  signatureMethod = DEFAULT_SIGNATURE_METHOD,
-  hashMethod = DEFAULT_HASH_METHODS,
-}: {
-  token: string;
-  secret: string;
-  issuer: string;
-  audience: string;
-  /** @default 60 */
-  leeway?: number;
-  /** @default 'HMAC' */
-  signatureMethod?: string;
-  /** @default 'SHA-256' */
-  hashMethod?: string;
-}) {
+export async function verifyJWT(options: VerifyJWTOptions) {
+  const {
+    token,
+    secret,
+    issuer,
+    audience,
+    leeway,
+    signatureMethod,
+    hashMethod,
+  } = {
+    leeway: 60,
+    signatureMethod: DEFAULT_SIGNATURE_METHOD,
+    hashMethod: DEFAULT_HASH_METHODS,
+    ...options,
+  };
+
   const [header, payload, signature] = String(token).split(".");
   const key = await importKey({
     secret,
